Share avatar user projection methods
