+++ app/src/ui/history/commit-list-item.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { Commit } from '../../models/commit'
 import { GitHubRepository } from '../../models/github-repository'
-import { IAvatarUser } from '../../models/avatar'
+import { IAvatarUser, getAvatarUsersForCommit } from '../../models/avatar'
 import { RichText } from '../lib/rich-text'
 import { RelativeTime } from '../relative-time'
 import { getDotComAPIEndpoint } from '../../lib/api'
@@ -9,9 +9,6 @@ import { clipboard } from 'electron'
 import { showContextualMenu, IMenuItem } from '../main-process-proxy'
 import { CommitAttribution } from '../lib/commit-attribution'
 import { IGitHubUser } from '../../lib/databases/github-user-database'
-import { CommitIdentity } from '../../models/commit-identity'
-import { GitAuthor } from '../../models/git-author'
-import { generateGravatarUrl } from '../../lib/gravatar'
 import { AvatarStack } from '../lib/avatar-stack'
 
 interface ICommitProps {
@@ -28,44 +25,6 @@ interface ICommitListItemState {
   readonly avatarUsers: ReadonlyArray<IAvatarUser>
 }
 
-function getAvatarUserFromAuthor(
-  gitHubUsers: Map<string, IGitHubUser> | null,
-  author: CommitIdentity | GitAuthor
-) {
-  const gitHubUser =
-    gitHubUsers === null
-      ? null
-      : gitHubUsers.get(author.email.toLowerCase()) || null
-
-  const avatarURL = gitHubUser
-    ? gitHubUser.avatarURL
-    : generateGravatarUrl(author.email)
-
-  return {
-    email: author.email,
-    name: author.name,
-    avatarURL,
-  }
-}
-
-function getAvatarUsers(
-  gitHubUsers: Map<string, IGitHubUser> | null,
-  commit: Commit
-) {
-  const avatarUsers = []
-
-  avatarUsers.push(getAvatarUserFromAuthor(gitHubUsers, commit.author))
-  avatarUsers.push(
-    ...commit.coAuthors.map(x => getAvatarUserFromAuthor(gitHubUsers, x))
-  )
-
-  if (!commit.authoredByCommitter) {
-    avatarUsers.push(getAvatarUserFromAuthor(gitHubUsers, commit.committer))
-  }
-
-  return avatarUsers
-}
-
 /** A component which displays a single commit in a commit list. */
 export class CommitListItem extends React.Component<
   ICommitProps,
@@ -75,14 +34,17 @@ export class CommitListItem extends React.Component<
     super(props)
 
     this.state = {
-      avatarUsers: getAvatarUsers(props.gitHubUsers, props.commit),
+      avatarUsers: getAvatarUsersForCommit(props.gitHubUsers, props.commit),
     }
   }
 
   public componentWillReceiveProps(nextProps: ICommitProps) {
     if (nextProps.commit !== this.props.commit) {
       this.setState({
-        avatarUsers: getAvatarUsers(nextProps.gitHubUsers, nextProps.commit),
+        avatarUsers: getAvatarUsersForCommit(
+          nextProps.gitHubUsers,
+          nextProps.commit
+        ),
       })
     }
   }
